+++ scripts/Publish-Extension.ps1
@@ -1,2 +1,0 @@
-# Publish extension
-Invoke-Expression "vsce publish 1.0.0"