+++ scripts/vscode-Install-Extension.ps1
@@ -1,2 +1,2 @@
 # Vscode install vsix
-Invoke-Expression "code --install-extension smart-formatter-1.0.0.vsix"
+Invoke-Expression "code --install-extension smart-formatter-1.1.0.vsix"