+++ src/Models/Constants.ts
@@ -1,0 +1,11 @@
+// SmartFormatter is two word
+export const GET_EACH_WORD_REGEX = /[A-Z][a-z]+|[a-z]+|[A-Z]+/g;
+
+// SmartFormatter is one word
+export const GET_EACH_CONNECTED_STRING_REGEX = /[a-zA-Z0-9]+/g;
+
+export const LINE_BREAK_REGEX = /.*[\r\n]+|.+$/g;
+
+export const WHITE_SPACE_REGEX = /^\s+/g;
+
+export const NEW_LINE_REGEX = /\r\n/g;