+++ src/Models/IConfiguration.ts
@@ -2,11 +2,11 @@ export default interface IConfiguration {
   editorContextMenu: boolean;
   stringFormatter: {
     enable: boolean;
-    toSingleLine: { joinString: string };
+    mergeLines: { joinString: string };
   };
   variableFormatter: {
     enable: boolean;
-    toVariableCustom: { capitalizeWords: boolean; joinString: string | null };
+    toCustomVariable: { isFirstCharUpper: boolean; joinString: string | null };
   };
   escapeFormatter: {
     enable: boolean;